use long

+++ src/main.cu
@@ -15,8 +15,8 @@ int main(int argc, char *argv[]) {
     Params params;
     parseCommandLine(argc, argv, params);
 
-    int N = params.N;
-    int Q = params.Q;
+    long N = params.N;
+    long Q = params.Q;
 
     // Allocate host memory
     float *h_documents = (float *)malloc(N * D * sizeof(float));
